backend/requirements.txt
Flask
flask-cors
requests
Pillow
numpy
shapely
geojson
